Split indentation from pretty printer
