+++ src/pretty_printer/indentation_options.ts
@@ -1,0 +1,4 @@
+export interface IIndentationOptions {
+  alignTryCatch?: boolean;
+  globalClassSkipFirst?: boolean;
+}
